fastapi_server/requirements.txt
# python-dotenv
# fastapi
# uvicorn

# langchain==0.0.352
# langchain-core==0.1.23
# langchain-community==0.0.20
# langsmith==0.0.87

# openai==0.28.1
# pydantic==1.10.13


fastapi
uvicorn
python-dotenv

langchain>=0.1.0
langchain-core>=0.1.23
langchain-community>=0.0.20
langchain-openai>=0.0.6

openai>=1.3.0
pydantic<2
